Moved file to correct dir, changed name
