+++ ThemeParkTycoonGame.Fancy/Controls/GuestinfoWindow.xaml
@@ -1,12 +1,12 @@
-﻿<Window x:Class="ThemeParkTycoonGame.Fancy.Guestinfo"
+﻿<Window x:Class="ThemeParkTycoonGame.Fancy.Controls.GuestinfoWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:ThemeParkTycoonGame.Fancy"
+        xmlns:local="clr-namespace:ThemeParkTycoonGame.Fancy.Controls"
         mc:Ignorable="d"
-        Title="Guestinfo" Height="450" Width="800">
-    <StackPanel>
-        <Label FontSize="20" FontWeight="Bold">Guestinfo:</Label>
-    </StackPanel>
+        Title="GuestinfoWindow" Height="450" Width="800">
+    <Grid>
+        
+    </Grid>
 </Window>
